+++ changelog.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - `idps` endpoint, view, and VAT
+- `returnees` table, view, and VAT
 - new constraint: `greater\_than\_constraint`
 
 ## 0.8.15
@@ -20,6 +21,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Changed
 
+- Added idps table and view (including materialised view)
 - Applied feedback from this
   [SQLAlchemy discussion](sqlalchemy/sqlalchemy#11748)
   to allow usage of type annotations defined in Base classes